test: migrate post_commit.rs unit tests to integration tests
Migrated all 12 tests from src/authorship/post_commit.rs to tests/integration/post_commit_unit.rs:
- 9 pure unit tests for count_line_ranges and should_skip_expensive_post_commit_stats
- 3 integration tests converted from TmpRepo to TestRepo

Made public in src/authorship/post_commit.rs:
- Constants: STATS_SKIP_MAX_HUNKS, STATS_SKIP_MAX_ADDED_LINES, STATS_SKIP_MAX_FILES_WITH_ADDITIONS, STATS_SKIP_MAX_DELETED_LINES
- Struct: StatsCostEstimate and its fields
- Functions: should_skip_expensive_post_commit_stats, count_line_ranges
